fix some statemachine problems
also figured out how to most easily set up generalized npc navigation in scenes with navigation regions. Navigating via tilemaps was really nice for the smoothness of the pathing when wandering, but it wasn't worth the hassle. Navigation via navigation regions has it's own quirks so far in testing, but I think that'll go away when levels are more populated with terrain because it's the big open space on the right that seems to be confusing pathing. We'll see how everything is once I re-implement slime chasing